backend/pyproject.toml
[tool.poetry]
name = "vestika"
version = "0.1.0"
description = "portfolio"
authors = ["Michael Palarya, Dan Sterenson, Ben Sterenson"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.12,<4"
fastapi = ">=0.115.14,<0.116.0"
yfinance = ">=0.2.64,<0.3.0"
pymaya = "0.5.1"
black = ">=25.1.0,<26.0.0"
pyyaml = ">=6.0.2,<7.0.0"
httpx = ">=0.28.1,<0.29.0"
pymongo = ">=4.8.0,<5.0.0"
redis = ">=6.2.0,<7.0.0"
loguru = ">=0.7.3,<0.8.0"
uvicorn = ">=0.34.3,<0.35.0"
fakeredis = "^2.30.1"
PyYAML = "^6.0"
python-multipart = "^0.0.20"
pydantic-settings = "^2.2.1"
firebase-admin = "^6.9.0"
finnhub-python = "^2.4.24"
google-genai = "^0.3.0"
gnews = "^0.3.9"
python-telegram-bot = "^21.6"
pynadlan = "^0.2.1"
apscheduler = "^3.10.4"
mixpanel = "^4.10.1"
aiohttp = "^3.13.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["slow: marks tests that hit real APIs (deselect with '-m \"not slow\"')"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
